refactor: extract sidebar navigation pieces
